pyproject.toml
[tool.poetry]
name = "search-for-vacancies"
version = "0.1.0"
description = ""
authors = ["kirill8769 <tkkrozn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"


[tool.poetry.group.lint.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.8.0"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = [".venv", "tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
